requirements.txt
# Clawbot v3 — Dependencies
python-telegram-bot>=21.0
python-dotenv
openai>=1.0
google-genai>=1.0
mss
Pillow
pynput
